requirements.txt
Django==5.1.2
python-decouple==3.8
psycopg2-binary==2.9.10
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
Unidecode==1.3.8